Eliminate package objects
