+++ src/test/scala/symsim/examples/concrete/simplemaze/MazeIsAgentSpec.scala
